pyproject.toml
[project]
name = "scanrole-api"
version = "0.1.0"
description = "Read-only API for ScanRole role explorer"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "httpx>=0.26",
    "pymysql>=1.1",
    "python-dotenv>=1.0",
    "pydantic>=2.5",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
pythonpath = ["src"]
